requirements.txt
streamlit>=1.24.0
ollama>=0.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.21.0
duckduckgo-search>=3.8.0
